// src/app/models/payment.ts
export class Payment {
    idPayment?: string;
    paymentDate: string;
    status: string
    paidMonth: string;
    paidYear: number
    amountPaid: number;
    paymentMethod: string;
    receiptNumber: string;
    notes?: string;
    createdAt?: string;
    updatedAt?: string
    idStudent: string;
}